% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvmCarregaArqAnualInformes.R
\name{cvmCarregaArqAnualInformes}
\alias{cvmCarregaArqAnualInformes}
\title{Funcao que carrega para um data.frame os informes diarios de fundos a partir de arquivo
zipado contendo os xmls referentes aos mesmos.}
\usage{
cvmCarregaArqAnualInformes(nomeArq, listaFundos = NULL)
}
\arguments{
\item{nomeArq}{nome do arquivo zipado que contem os xml relativos aos informes diarios.}

\item{listaFundos}{vetor de caracters contendo os CNPJ dos fundos dos quais deseja-se extrair as informacoes}
}
\value{
objeto do tipo data.frame contendo os informes diarios dos fundos.
}
\description{
Funcao que carrega para um data.frame os informes diarios de fundos a partir de arquivo
zipado contendo os xmls referentes aos mesmos.
}
\details{
Esta funcao carrega para um data.frame os informes diarios de fundos a partir de arquivo
zipado contendo os xmls referentes aos mesmos.
}
\examples{
\dontrun{
dfInformes <- cvmCarregaArqAnualInformes(nomeArq, listaFundos)
}
}
\author{
Bruno M. S. S. Melo <ctbrunomelo@gmail.com>
}
\seealso{
\url{http://sistemas.cvm.gov.br}
}

