% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvmDownload.R
\name{cvmDownload}
\alias{cvmDownload}
\title{Funcao que realiza o download de algum arquivo que tenha sido previamente liberado pelo
webservice da CVM.}
\usage{
cvmDownload(autorizDownload, nomeArq = "cvm.zip")
}
\arguments{
\item{autorizDownload}{data.frame retornado por alguma funcao do tipo cvmSolicAUtorizDownload*}

\item{nomeArq}{nome do arquivo de destino}
}
\description{
Funcao que realiza o download de algum arquivo que tenha sido previamente liberado pelo
webservice da CVM.
}
\details{
Esta funcao realiza download de algum arquivo que ja tenha sido liberado pela CVM por meio
de alguma funcao do tipo "cvmSolicAUtorizDownload*".
}
\examples{
\dontrun{
cvmDownload(autorizDownload)
}
}
\author{
Bruno M. S. S. Melo <ctbrunomelo@gmail.com>
}
\seealso{
\url{http://sistemas.cvm.gov.br}
}

