% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvmLogin.R
\name{cvmLogin}
\alias{cvmLogin}
\title{Efetua o Login no webservice da CVM}
\usage{
cvmLogin(iNrSist, strSenha, verSoap = "1.1")
}
\arguments{
\item{identificador}{do sistema junto a CVM}

\item{ver}{versao SOAP ("1.1" ou "1.2")}
}
\value{
um objeto do tipo data.frame contendo o GUID e idSessao.
}
\description{
Efetua o Login no webservice da CVM
}
\details{
Esta funcao efetua o Login no webservice retornando o GUID e idSessao a serem utilizados
em outros métodos disponibilizados pela CVM.
Obs.: Para informacoes relativas ao cadastro, consulte o site da CVM.
}
\examples{
\dontrun{
dfSession <- cvmLogin(iNrSist, strSenha)
}
}
\author{
Bruno M. S. S. Melo
}
\seealso{
\url{http://sistemas.cvm.gov.br}
}

