% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvmSolicAutorizDownloadCadastro.R
\name{cvmSolicAutorizDownloadCadastro}
\alias{cvmSolicAutorizDownloadCadastro}
\title{Solicita autorizacao para o download do cadastro dos fundos registrados na CVM.}
\usage{
cvmSolicAutorizDownloadCadastro(loginInfo, refDate)
}
\arguments{
\item{loginInfo}{data.frame retornado pela funcao cvmLogin, contendo informacoes de login.}

\item{refDate}{string contendo data de referencia no formato aaaa-mm-dd.}
}
\value{
um objeto do tipo data.frame contendo a url para efetuar o download solicitado.
}
\description{
Solicita autorizacao para o download do cadastro dos fundos registrados na CVM.
}
\details{
Esta funcao solicita autorizacao pala efetuar o download do cadastro dos fundos registrados na CVM.
}
\examples{
\dontrun{
dfAutorizacao <- cvmSolicAutorizDownloadCadastro(loginInfo = dfLogin)
}
}
\author{
Bruno M. S. S. Melo <ctbrunomelo@gmail.com>
}
\seealso{
\url{http://sistemas.cvm.gov.br}
}

