% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvmRetornaCorpoSoap.R
\name{cvmRetornaCorpoSoap}
\alias{cvmRetornaCorpoSoap}
\title{Retorna o formato da mensagem SOAP a ser enviado ao webservice da CVM}
\usage{
cvmRetornaCorpoSoap(op, ver)
}
\arguments{
\item{op}{nome do metodo desejado}

\item{ver}{versao SOAP}
}
\value{
um objeto do tipo data.frame contendo o resumo da politica de investimentos do RPPS.
}
\description{
Retorna o formato da mensagem SOAP a ser enviado ao webservice da CVM
}
\details{
Esta funcao extrai o corpo da mensagem SOAP a ser enviada ao webservice da CVM, de acordo
com o nome do método fornecido pelo parametro ("op") e versao ("ver") SOAP desejada.
}
\examples{
\dontrun{
body <- cvmRetornaCorpoSoap("Login", "1.1")
}
}
\author{
Bruno M. S. S Melo
}
\seealso{
\code{RCurl}, \code{XML} \code{stringr}
}

